Order Class: - Add passenger to order. - Remove passenger from order.
Order Repository:
- ID generation (incremental).
- Add passenger to order implementation.
- Remove passenger from order implementation.

Order Singleton:
- Get orders by agent agent code.
- Get orders by passenger ID.
- Delete order by order ID.
- Cancel order by order ID.
- Reactivate order by order ID.
- Save order data to Json file.

+++ src/model/objects/Agent.java
@@ -13,7 +13,7 @@ import java.util.Objects;
 public class Agent extends Person {
 
     private String userName, password;
-    private int agentCode;
+    private long agentCode;
     private boolean isVerified;
     private int permissionLevel;
 
@@ -45,8 +45,8 @@ public class Agent extends Person {
     public void setVerified(boolean verified) {  isVerified = verified;}
 
     // Getters
-    public String getUserName() {  return this.userName;}
-    public int getAgentCode() {return this.agentCode;}
+    public String getUserName() { return this.userName; }
+    public long getAgentCode() { return this.agentCode; }
     public boolean isVerified() { return this.isVerified; }
     public int getPermissionLevel() { return this.permissionLevel; }
     public String getPassword() { return this.password; }
